get api user info from secret
