+++ deploy/secrets.yaml
@@ -5,8 +5,8 @@ metadata:
   namespace: argocd-interlace
 type: Opaque    
 data:
-  cosign.key: UkVQTEFDRSBUSElTCg==
-  pubring.gpg: UkVQTEFDRSBUSElTCg==
+  signKey: UkVQTEFDRSBUSElTCg==
+  verifyKey: UkVQTEFDRSBUSElTCg==
 ---
 apiVersion: v1
 kind: Secret
@@ -15,7 +15,7 @@ metadata:
   namespace: argocd-interlace
 type: Opaque    
 data:
-  ARGOCD_NAMESPACE: UkVQTEFDRSBUSElTCg==
-  ARGOCD_USER: UkVQTEFDRSBUSElTCg==
-  ARGOCD_USER_PWD: UkVQTEFDRSBUSElTCg==
+  argocdNamespace: UkVQTEFDRSBUSElTCg==
+  argocdAPIUsername: UkVQTEFDRSBUSElTCg==
+  argocdAPIPassword: UkVQTEFDRSBUSElTCg==
