+++ pkg/utils/utils.go
@@ -18,22 +18,13 @@ package utils
 
 import (
 	"bytes"
-	"context"
 	"crypto/sha256"
-	"crypto/tls"
-	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 
-	"github.com/argoproj-labs/argocd-interlace/pkg/config"
-	"github.com/argoproj/argo-cd/v2/pkg/apiclient"
-	"github.com/argoproj/argo-cd/v2/pkg/apiclient/application"
-	argoio "github.com/argoproj/argo-cd/v2/util/io"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"k8s.io/client-go/kubernetes"
@@ -48,17 +39,14 @@ const (
 	PROVENANCE_FILE_NAME      = "provenance.yaml"
 	ATTESTATION_FILE_NAME     = "attestation.json"
 	TMP_DIR                   = "/tmp/output"
-	PRIVATE_KEY_PATH          = "/etc/keys/cosign.key"
-	KEYRING_PUB_KEY_PATH      = "/etc/keys/pubring.gpg"
+	PRIVATE_KEY_PATH          = "/etc/keys/signKey"
+	KEYRING_PUB_KEY_PATH      = "/etc/keys/verifyKey"
 	SIG_ANNOTATION_NAME       = "cosign.sigstore.dev/signature"
 	MSG_ANNOTATION_NAME       = "cosign.sigstore.dev/message"
-	RETRY_ATTEMPTS            = 10
 )
 
-const patchType = "application/merge-patch+json"
-
-//GetClient returns a kubernetes client
-func GetClient(configpath string) (*kubernetes.Clientset, *rest.Config, error) {
+//GetK8sClient returns a kubernetes client and config
+func GetK8sClient(configpath string) (*kubernetes.Clientset, *rest.Config, error) {
 
 	if configpath == "" {
 		log.Debug("Using Incluster configuration")
@@ -108,157 +96,6 @@ func WriteToFile(str, dirPath, filename string) error {
 
 	return nil
 
-}
-
-func QueryAPI(url, requestType, bearerToken string, data map[string]interface{}) (string, error) {
-
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-
-	bearer := ""
-	if bearerToken != "" {
-		bearer = "Bearer " + bearerToken
-	}
-
-	var dataJson []byte
-	if data != nil {
-		dataJson, _ = json.Marshal(data)
-	} else {
-		dataJson = nil
-	}
-	req, err := http.NewRequest(requestType, url, bytes.NewBuffer(dataJson))
-	if err != nil {
-		log.Errorf("Error %s ", err.Error())
-		return "", err
-	}
-
-	if bearer != "" {
-		req.Header.Add("Authorization", bearer)
-	}
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Errorf("Error %s", err.Error())
-		return "", err
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Errorf("Error %s ", err.Error())
-		return "", err
-	}
-
-	return string([]byte(body)), nil
-}
-
-func RetriveDesiredManifest(appName string) ([]string, error) {
-	interlaceConfig, err := config.GetInterlaceConfig()
-	if err != nil {
-		log.Errorf("Error in loading config: %s", err.Error())
-	}
-
-	baseUrl := interlaceConfig.ArgocdApiBaseUrl
-	manifestsAPIURL := fmt.Sprintf("%s/api/v1/applications/%s/manifests", baseUrl, appName)
-
-	token, err := GetArgoCDUserToken(interlaceConfig.ArgocdApiBaseUrl, interlaceConfig.ArgocdUser, interlaceConfig.ArgocdUserPwd)
-	if err != nil {
-		return nil, errors.Wrap(err, "error when getting argocd user token")
-	}
-	manifestsResponse, err := QueryAPI(manifestsAPIURL, "GET", token, nil)
-	if err != nil {
-		return nil, errors.Wrap(err, "error occured while querying argocd REST API")
-	}
-
-	var respMap map[string]interface{}
-	err = json.Unmarshal([]byte(manifestsResponse), &respMap)
-	if err != nil {
-		return nil, errors.Wrap(err, "error occured while marshaling the manifest API response")
-	}
-
-	var manifestsIf interface{}
-	var manifestsIfList []interface{}
-	var manifests []string
-	var ok bool
-	if manifestsIf, ok = respMap["manifests"]; !ok {
-		return nil, errors.New("`manifests` is not found in the manifest API response")
-	}
-	if manifestsIfList, ok = manifestsIf.([]interface{}); !ok {
-		return nil, fmt.Errorf("expect `[]interface{}` for manifests, but got `%T`", manifestsIf)
-	}
-	for _, manifestIf := range manifestsIfList {
-		if tmpMnf, ok := manifestIf.(string); !ok {
-			log.Errorf("expect `string` in the manifest list, but got `%T`", manifestIf)
-		} else {
-			manifests = append(manifests, tmpMnf)
-		}
-	}
-	return manifests, nil
-}
-
-func GetArgoCDUserToken(apiBaseURL, argocdUser, argocdUserPass string) (string, error) {
-	sessionURL := fmt.Sprintf("%s/api/v1/session", apiBaseURL)
-	input := map[string]interface{}{
-		"username": argocdUser,
-		"password": argocdUserPass,
-	}
-	sessiondata, err := QueryAPI(sessionURL, "POST", "", input)
-	if err != nil {
-		return "", err
-	}
-	log.Debugf("sessiondata from argocd api: %s", string(sessiondata))
-	var output map[string]interface{}
-	err = json.Unmarshal([]byte(sessiondata), &output)
-	if err != nil {
-		return "", err
-	}
-	sessionTokenIf, ok := output["token"]
-	if !ok {
-		return "", errors.New("argocd user token is not found in the response from the session api")
-	}
-	sessionToken, ok := sessionTokenIf.(string)
-	if !ok {
-		return "", fmt.Errorf("the token returned from the session api was not a string, but %T", sessionTokenIf)
-	}
-	return sessionToken, nil
-}
-
-func PatchResource(appName, namespace, resourceName, group, version, kind string, patchBytes []byte) error {
-	interlaceConfig, err := config.GetInterlaceConfig()
-	if err != nil {
-		return errors.Wrap(err, "failed to load interlace config")
-	}
-	baseUrl := interlaceConfig.ArgocdApiBaseUrl
-	opt := &apiclient.ClientOptions{
-		ServerAddr: baseUrl,
-		Insecure:   true,
-		GRPCWeb:    true,
-	}
-	client, err := apiclient.NewClient(opt)
-	if err != nil {
-		return errors.Wrap(err, "failed to initialize argocd client set")
-	}
-	conn, appClient, err := client.NewApplicationClient()
-	if err != nil {
-		return errors.Wrap(err, "failed to initialize argocd application client")
-	}
-	defer argoio.Close(conn)
-
-	resourcePatchRequest := &application.ApplicationResourcePatchRequest{
-		Name:         &appName,
-		Namespace:    namespace,
-		ResourceName: resourceName,
-		Group:        group,
-		Version:      version,
-		Kind:         kind,
-		Patch:        string(patchBytes),
-		PatchType:    patchType,
-	}
-	resourcePatchResp, err := appClient.PatchResource(context.Background(), resourcePatchRequest)
-	if err != nil {
-		return errors.Wrap(err, "failed to patch resource")
-	}
-	log.Debugf("patch resource response: %s", resourcePatchResp.Manifest)
-	return nil
 }
 
 func FileExist(fpath string) bool {
