+++ scripts/setup.sh
@@ -3,8 +3,8 @@
 set -e
 
 ARGOCD_NAMESPACE=$1
-ARGOCD_ADMIN_USERNAME=$2
-ARGOCD_ADMIN_PASSWORD=$3
+ARGOCD_API_USERNAME=$2
+ARGOCD_API_PASSWORD=$3
 SIGN_KEY_PATH=$4
 VERIFY_KEY_PATH=$5
 
@@ -21,8 +21,8 @@ if ! [ -x "$(command -v kubectl)" ]; then
 fi
 
 # configure argocd API token to `argocd-config-secret`
-kubectl patch secret argocd-config-secret -n argocd-interlace -p="{\"data\":{\"ARGOCD_NAMESPACE\":\"$(echo -n $ARGOCD_NAMESPACE | base64)\",\"ARGOCD_USER\":\"$(echo -n $ARGOCD_ADMIN_USERNAME | base64)\",\"ARGOCD_USER_PWD\":\"$(echo -n $ARGOCD_ADMIN_PASSWORD | base64)\"}}"
+kubectl patch secret argocd-config-secret -n argocd-interlace -p="{\"data\":{\"argocdNamespace\":\"$(echo -n $ARGOCD_NAMESPACE | base64)\",\"argocdAPIUser\":\"$(echo -n $ARGOCD_API_USERNAME | base64)\",\"argocdAPIPassword\":\"$(echo -n $ARGOCD_API_PASSWORD | base64)\"}}"
 
 # configure `argocd-interlace-keys`
-kubectl patch secret argocd-interlace-keys -n argocd-interlace -p="{\"data\":{\"cosign.key\":\"$(cat $SIGN_KEY_PATH | base64)\",\"pubring.gpg\":\"$(cat $VERIFY_KEY_PATH | base64)\"}}"
+kubectl patch secret argocd-interlace-keys -n argocd-interlace -p="{\"data\":{\"signKey\":\"$(cat $SIGN_KEY_PATH | base64)\",\"verifyKey\":\"$(cat $VERIFY_KEY_PATH | base64)\"}}"
